Add Lumen FormRequest compatibility
