+++ src/Http/LumenFormRequest.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Dingo\Api\Http;
+
+use Illuminate\Contracts\Validation\ValidatesWhenResolved;
+use Laravel\Lumen\Http\Request as LumenRequest;
+
+class LumenFormRequest extends LumenRequest implements ValidatesWhenResolved
+{
+    use RequestTrait;
+}
